requirements.txt
Flask
Flask-Bootstrap
Flask-Testing
Flask-WTF
redis
rq
prometheus-client
prometheus_flask_exporter
sentry-sdk[flask]
sentry-sdk[rq]
sentry-sdk[logging]
nomics-python
python-dotenv
requests
mysql-connector-python
ccxt
pandas